Bug 27990: Replace obsolete title-string sorting: Patrons
This patch modifies several patron templates to replace the use of
the "title-string" DataTables sorting method with the newer "data-order"
attribute.

To test, apply the patch and view a patron record in the staff
interface. Check the following pages to confirm that columns containing
dates sort correctly when using any setting of the "dateformat" system
preference:

- Details: DataTables initialization removed because it is obsolete.
  Confirm that table of holds in the holds tab is still a working
  DataTable.
- Accounting
- Accounting -> Transactions,
- Accounting -> Transactions -> Details of a fee (DataTables-related
  markup removed because there are no DataTables).
- Circulation history
- Holds history
- Notices

Signed-off-by: Amit Gupta <amitddng135@gmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>