Bug 28035: Improve breadcrumbs of cataloging search page
This patch expands the logic around the cataloging search page's
breadcrumbs so that more context is provided in the breadcrumbs menu.

The page title should also reflect this change.

Some markup in the body of the page has been updated to clarify the page
structure using headings.

To test, apply the patch and go to Cataloging.

 - The page title and breadcrumbs menu should show "Home > Cataloging"
 - Perform a search which will return results from the catalog, from the
   reservoir, or both. The page title and breadcrumbs menu should show:
   "Home > Cataloging > Search results."
 - Perform a search which will return no results. The page
   title and breadcrumbs should be: "Home > Cataloging > No results
   found."

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>