Bug 28086: OpacMaintenance page should use ReplytoDefault if set
This patch changes the OPAC maintenance template so that it checks for a
value in the ReplyToDefault system preference when displaying the
"please contact the site administrator" email link.

Also changed: If neither ReplytoDefault nor KohaAdminEmailAddress is
defined, it won't show the link at all.

To test, apply the patch and set the OpacMaintenance preference to
"Show."

 - In the OPAC you should see the "System Maintenance" page.
 - Test this page with various settings:
   - ReplytoDefault and KohaAdminEmailAddress both populated.
   - Only ReplytoDefault
   - Only KohaAdminEmailAddress
   - Neither.

 Confirm that the correct email address is used in each case.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>