Bug 28055: (follow-up) Revert changes to aoColumnDefs option name
This patch undoes changes to two lines in columns_settings.inc which
were triggering unwanted side-effects. There are several templates and
JavaScript files which use "aoColumnDefs" to pass options to columns
settings, and if we change this file without changing those at the same
time things will break.

When these lines are updated it must be done so at the same time as we
modify the templates which pass that same option.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>