Bug 27830: (follow-up) Merge DE and FR address formats
Library addresses don't include the use of "Street Type" and "Street
Number." DE and FR address formats differ only in the position of street
type and number in the address sequence.

This patch merges DE and FR address markup and uses the unified block if
the address format system preference ISN'T "us".

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>