Bug 27924: Display number of holds awaiting pickup on check out screens
This patch modifies the checkout page so that waiting holds are
displayed separately depending on whether they are waiting at the
current branch or not. A count of each number of waiting holds is
displayed too.

Unrelated change: A missing </li> has been added for markup validity.

To test, apply the patch and view the checkout screen in these
situations:

 - A patron with no waiting holds.
 - A patron with one or more holds waiting at the current library.
 - A patron with one or more holds waiting at another library.
 - A patron with holds waiting at both the current library and at other
   libraries.

In each case, the display of waiting hold information should be correct,
including the count of holds of each kind.

Signed-off-by: Barbara Johnson <barbara.johnson@bedfordtx.gov>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>